details screen created & rest all updated

+++ lib/models/movie.dart
@@ -25,7 +25,7 @@ class Movie{
         overview: json["overview"],
         posterPath: json["poster_path"],
         releaseDate: json["release_date"],
-        voteAverage: json["vote_average"]
+        voteAverage: json["vote_average"].toDouble(),
     );
   }
